Move interfaces into Paseto::Interface
